requirements.txt
einops==0.4.1
numpy==1.20.3
opencv_python_headless==4.5.5.64
Pillow==8.4.0
PyYAML==6.0
setuptools==58.0.4
tensorboardX==2.5.1
torch==1.11.0
torchmetrics==1.0.1
torchvision==0.12.0
tqdm==4.62.3
